+++ src/main/java/net/roguelogix/biggerreactors/classic/reactor/ReactorMultiblockController.java
@@ -246,7 +246,7 @@ public class ReactorMultiblockController extends RectangularMultiblockController
     public void tick() {
         
         simulation.tick();
-        if (!Float.isNaN(simulation.FEProducedLastTick) && simulation.isPassive()) {
+        if (!Double.isNaN(simulation.FEProducedLastTick) && simulation.isPassive()) {
             storedPower += simulation.FEProducedLastTick;
             if (storedPower > Config.Reactor.PassiveBatterySize) {
                 storedPower = Config.Reactor.PassiveBatterySize;
