+++ src/main/java/net/roguelogix/biggerreactors/classic/reactor/simulation/ClassicReactorSimulation.java
@@ -19,35 +19,35 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
             new Vector2i(0, 1),
             new Vector2i(0, -1),
     };
-    public float fuelConsumedLastTick = 0f;
-    public float FEProducedLastTick = 0f;
+    public double fuelConsumedLastTick = 0f;
+    public double FEProducedLastTick = 0f;
     private int x, y, z;
     private ReactorModeratorRegistry.ModeratorProperties[][][] moderatorProperties;
     private ControlRod[][] controlRodsXZ;
-    private float fuelToReactorHeatTransferCoefficient = 0;
-    private float reactorToCoolantSystemHeatTransferCoefficient = 0;
-    private float reactorHeatLossCoefficient = 0;
+    private double fuelToReactorHeatTransferCoefficient = 0;
+    private double reactorToCoolantSystemHeatTransferCoefficient = 0;
+    private double reactorHeatLossCoefficient = 0;
     private boolean active = false;
-    private float fuelFertility = 1f;
-    private float fuelHeat = Config.Reactor.AmbientTemperature;
-    private float reactorHeat = Config.Reactor.AmbientTemperature;
+    private double fuelFertility = 1f;
+    private double fuelHeat = Config.Reactor.AmbientTemperature;
+    private double reactorHeat = Config.Reactor.AmbientTemperature;
     private boolean passive = true;
     private int rodToIrradiate = 0;
     private int yLevelToIrradiate = 0;
     
-    public static float getRFFromVolumeAndTemp(float volume, float temperature) {
+    public static double getRFFromVolumeAndTemp(double volume, double temperature) {
         return temperature * volume * Config.Reactor.FEPerCentigradePerUnitVolume;
     }
     
-    public static float getTempFromVolumeAndRF(float volume, float rf) {
+    public static double getTempFromVolumeAndRF(double volume, double rf) {
         return rf / (volume * Config.Reactor.FEPerCentigradePerUnitVolume);
     }
     
-    float reactorVolume() {
+    double reactorVolume() {
         return x * y * z;
     }
     
-    float fuelRodVolume() {
+    double fuelRodVolume() {
         return controlRods.size() * y;
     }
     
@@ -83,7 +83,7 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         return controlRods.size();
     }
     
-    public void setControlRodInsertion(int x, int z, float insertion) {
+    public void setControlRodInsertion(int x, int z, double insertion) {
         controlRodsXZ[x][z].insertion = insertion;
     }
     
@@ -135,7 +135,7 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         
         {
             // decay fertility, RadiationHelper.tick in old BR, this is copied, mostly
-            float denominator = Config.Reactor.FuelFertilityDecayDenominator;
+            double denominator = Config.Reactor.FuelFertilityDecayDenominator;
             if (!active) {
                 // Much slower decay when off
                 denominator *= Config.Reactor.FuelFertilityDecayDenominatorInactiveMultiplier;
@@ -146,16 +146,16 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         }
         
         // Heat Transfer: Fuel Pool <> Reactor Environment
-        float tempDiff = fuelHeat - reactorHeat;
+        double tempDiff = fuelHeat - reactorHeat;
         if (tempDiff > 0.01f) {
-            float rfTransferred = tempDiff * fuelToReactorHeatTransferCoefficient;
-            float fuelRf = getRFFromVolumeAndTemp(fuelRodVolume(), fuelHeat);
+            double rfTransferred = tempDiff * fuelToReactorHeatTransferCoefficient;
+            double fuelRf = getRFFromVolumeAndTemp(fuelRodVolume(), fuelHeat);
             
             fuelRf -= rfTransferred;
             fuelHeat = getTempFromVolumeAndRF(fuelRodVolume(), fuelRf);
             
             // Now see how much the reactor's temp has increased
-            float reactorRf = getRFFromVolumeAndTemp(reactorVolume(), reactorHeat);
+            double reactorRf = getRFFromVolumeAndTemp(reactorVolume(), reactorHeat);
             reactorRf += rfTransferred;
             reactorHeat = getTempFromVolumeAndRF(reactorVolume(), reactorRf);
         }
@@ -163,8 +163,8 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         // If we have a temperature differential between environment and coolant system, move heat between them.
         tempDiff = reactorHeat - getCoolantTemperature();
         if (tempDiff > 0.01f) {
-            float rfTransferred = tempDiff * reactorToCoolantSystemHeatTransferCoefficient;
-            float reactorRf = getRFFromVolumeAndTemp(reactorVolume(), reactorHeat);
+            double rfTransferred = tempDiff * reactorToCoolantSystemHeatTransferCoefficient;
+            double reactorRf = getRFFromVolumeAndTemp(reactorVolume(), reactorHeat);
             
             if (passive) {
                 rfTransferred *= Config.Reactor.PassiveCoolingTransferEfficiency;
@@ -181,8 +181,8 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         // Do passive heat loss - this is always versus external environment
         tempDiff = reactorHeat - Config.Reactor.AmbientTemperature;
         if (tempDiff > 0.000001f) {
-            float rfLost = Math.max(1f, tempDiff * reactorHeatLossCoefficient); // Lose at least 1RF/t
-            float reactorNewRf = Math.max(0f, getRFFromVolumeAndTemp(reactorVolume(), reactorHeat) - rfLost);
+            double rfLost = Math.max(1f, tempDiff * reactorHeatLossCoefficient); // Lose at least 1RF/t
+            double reactorNewRf = Math.max(0f, getRFFromVolumeAndTemp(reactorVolume(), reactorHeat) - rfLost);
             reactorHeat = getTempFromVolumeAndRF(reactorVolume(), reactorNewRf);
         }
         
@@ -199,7 +199,7 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         passive = passivelyCooled;
     }
     
-    private float getCoolantTemperature() {
+    private double getCoolantTemperature() {
         if (passive) {
             return Config.Reactor.AmbientTemperature;
         }
@@ -242,42 +242,42 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         long baseFuelAmount = fuelTank.getFuelAmount() + (fuelTank.getWasteAmount() / 100);
         
         // Intensity = how strong the radiation is, hardness = how energetic the radiation is (penetration)
-        float rawRadIntensity = (float) baseFuelAmount * Config.Reactor.FissionEventsPerFuelUnit;
+        double rawRadIntensity = (double) baseFuelAmount * Config.Reactor.FissionEventsPerFuelUnit;
         
         // Scale up the "effective" intensity of radiation, to provide an incentive for bigger reactors in general.
-        float scaledRadIntensity = (float) Math.pow((rawRadIntensity), Config.Reactor.FuelReactivity);
+        double scaledRadIntensity = Math.pow((rawRadIntensity), Config.Reactor.FuelReactivity);
         
         // Scale up a second time based on scaled amount in each fuel rod. Provides an incentive for making reactors that aren't just pancakes.
-        scaledRadIntensity = (float) Math.pow((scaledRadIntensity / controlRods.size()), Config.Reactor.FuelReactivity) * controlRods.size();
+        scaledRadIntensity = Math.pow((scaledRadIntensity / controlRods.size()), Config.Reactor.FuelReactivity) * controlRods.size();
         
         // Apply control rod moderation of radiation to the quantity of produced radiation. 100% insertion = 100% reduction.
-        float controlRodModifier = (100 - controlRods.get(rodToIrradiate).insertion) / 100f;
+        double controlRodModifier = (100 - controlRods.get(rodToIrradiate).insertion) / 100f;
         scaledRadIntensity = scaledRadIntensity * controlRodModifier;
         rawRadIntensity = rawRadIntensity * controlRodModifier;
         
         // Now nerf actual radiation production based on heat.
-        float effectiveRadIntensity = scaledRadIntensity * (1f + (float) (-Config.Reactor.RadIntensityScalingMultiplier * Math.exp(-10f * Config.Reactor.RadIntensityScalingShiftMultiplier * Math.exp(-0.001f * Config.Reactor.RadIntensityScalingRateExponentMultiplier * fuelHeat))));
+        double effectiveRadIntensity = scaledRadIntensity * (1f + (double) (-Config.Reactor.RadIntensityScalingMultiplier * Math.exp(-10f * Config.Reactor.RadIntensityScalingShiftMultiplier * Math.exp(-0.001f * Config.Reactor.RadIntensityScalingRateExponentMultiplier * fuelHeat))));
         
         // Radiation hardness starts at 20% and asymptotically approaches 100% as heat rises.
         // This will make radiation harder and harder to capture.
-        float radHardness = 0.2f + (float) (0.8 * radiationPenaltyBase);
+        double radHardness = 0.2f + (double) (0.8 * radiationPenaltyBase);
         
         // Calculate based on propagation-to-self
-        float rawFuelUsage = (Config.Reactor.FuelPerRadiationUnit * rawRadIntensity / getFertility()) * Config.Reactor.FuelUsageMultiplier; // Not a typo. Fuel usage is thus penalized at high heats.
-        float fuelRfChange = Config.Reactor.FEPerRadiationUnit * effectiveRadIntensity;
-        float environmentRfChange = 0f;
+        double rawFuelUsage = (Config.Reactor.FuelPerRadiationUnit * rawRadIntensity / getFertility()) * Config.Reactor.FuelUsageMultiplier; // Not a typo. Fuel usage is thus penalized at high heats.
+        double fuelRfChange = Config.Reactor.FEPerRadiationUnit * effectiveRadIntensity;
+        double environmentRfChange = 0f;
         
         effectiveRadIntensity *= 0.25f; // We're going to do this four times, no need to repeat
         
-        float fuelAbsorbedRadiation = 0f;
+        double fuelAbsorbedRadiation = 0f;
         
         Vector2i position = new Vector2i();
         
         for (Vector2i direction : directions) {
             position.set(controlRods.get(rodToIrradiate).x, controlRods.get(rodToIrradiate).z);
             
-            float hardness = radHardness;
-            float intensity = effectiveRadIntensity;
+            double hardness = radHardness;
+            double intensity = effectiveRadIntensity;
             
             for (int i = 0; i < Config.Reactor.IrradiationDistance && intensity > 0.0001f; i++) {
                 position.add(direction);
@@ -290,7 +290,7 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
                 if (properties != null) {
                     // this is the simple part, which is why its first
                     
-                    float radiationAbsorbed = intensity * properties.absorption * (1f - hardness);
+                    double radiationAbsorbed = intensity * properties.absorption * (1f - hardness);
                     intensity = Math.max(0f, intensity - radiationAbsorbed);
                     hardness /= properties.moderation;
                     environmentRfChange += properties.heatEfficiency * radiationAbsorbed * Config.Reactor.FEPerRadiationUnit;
@@ -299,25 +299,25 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
                     // oh, oh ok, its a fuel rod
                     
                     // Scale control rod insertion 0..1
-                    float controlRodInsertion = Math.min(1f, Math.max(0f, controlRodsXZ[position.x][position.y].insertion / 100f));
+                    double controlRodInsertion = Math.min(1f, Math.max(0f, controlRodsXZ[position.x][position.y].insertion / 100f));
                     
                     // Fuel absorptiveness is determined by control rod + a heat modifier.
                     // Starts at 1 and decays towards 0.05, reaching 0.6 at 1000 and just under 0.2 at 2000. Inflection point at about 500-600.
                     // Harder radiation makes absorption more difficult.
-                    float baseAbsorption = (float) (1.0 - (Config.Reactor.FuelAbsorptionScalingMultiplier * Math.exp(-10 * Config.Reactor.FuelAbsorptionScalingShiftMultiplier * Math.exp(-0.001 * Config.Reactor.FuelAbsorptionScalingRateExponentMultiplier * fuelHeat)))) * (1f - (hardness / Config.Reactor.FuelHardnessDivisor));
+                    double baseAbsorption = (1.0 - (Config.Reactor.FuelAbsorptionScalingMultiplier * Math.exp(-10 * Config.Reactor.FuelAbsorptionScalingShiftMultiplier * Math.exp(-0.001 * Config.Reactor.FuelAbsorptionScalingRateExponentMultiplier * fuelHeat)))) * (1f - (hardness / Config.Reactor.FuelHardnessDivisor));
                     
                     // Some fuels are better at absorbing radiation than others
-                    float scaledAbsorption = Math.min(1f, baseAbsorption * Config.Reactor.FuelAbsorptionCoefficient);
+                    double scaledAbsorption = Math.min(1f, baseAbsorption * Config.Reactor.FuelAbsorptionCoefficient);
                     
                     // Control rods increase total neutron absorption, but decrease the total neutrons which fertilize the fuel
                     // Absorb up to 50% better with control rods inserted.
-                    float controlRodBonus = (1f - scaledAbsorption) * controlRodInsertion * 0.5f;
-                    float controlRodPenalty = scaledAbsorption * controlRodInsertion * 0.5f;
+                    double controlRodBonus = (1f - scaledAbsorption) * controlRodInsertion * 0.5f;
+                    double controlRodPenalty = scaledAbsorption * controlRodInsertion * 0.5f;
                     
-                    float radiationAbsorbed = (scaledAbsorption + controlRodBonus) * intensity;
-                    float fertilityAbsorbed = (scaledAbsorption - controlRodPenalty) * intensity;
+                    double radiationAbsorbed = (scaledAbsorption + controlRodBonus) * intensity;
+                    double fertilityAbsorbed = (scaledAbsorption - controlRodPenalty) * intensity;
                     
-                    float fuelModerationFactor = Config.Reactor.FuelModerationFactor;
+                    double fuelModerationFactor = Config.Reactor.FuelModerationFactor;
                     fuelModerationFactor += fuelModerationFactor * controlRodInsertion + controlRodInsertion; // Full insertion doubles the moderation factor of the fuel as well as adding its own level
                     
                     intensity = Math.max(0f, intensity - radiationAbsorbed);
@@ -339,8 +339,8 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         fuelConsumedLastTick = rawFuelUsage;
     }
     
-    protected void addReactorHeat(float newCasingHeat) {
-        if (Float.isNaN(newCasingHeat)) {
+    protected void addReactorHeat(double newCasingHeat) {
+        if (Double.isNaN(newCasingHeat)) {
             return;
         }
         
@@ -351,8 +351,8 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         }
     }
     
-    protected void addFuelHeat(float additionalHeat) {
-        if (Float.isNaN(additionalHeat)) {
+    protected void addFuelHeat(double additionalHeat) {
+        if (Double.isNaN(additionalHeat)) {
             return;
         }
         
@@ -363,27 +363,27 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         }
     }
     
-    public float getFertility() {
+    public double getFertility() {
         if (fuelFertility <= 1f) {
             return 1f;
         } else {
-            return (float) (Math.log10(fuelFertility) + 1);
+            return (double) (Math.log10(fuelFertility) + 1);
         }
     }
     
-    public float getFuelHeat() {
+    public double getFuelHeat() {
         return fuelHeat;
     }
     
-    public float getReactorHeat() {
+    public double getReactorHeat() {
         return reactorHeat;
     }
     
-    public float getFuelConsumedLastTick() {
+    public double getFuelConsumedLastTick() {
         return fuelConsumedLastTick;
     }
     
-    public float getFEProducedLastTick() {
+    public double getFEProducedLastTick() {
         return FEProducedLastTick;
     }
     
@@ -392,9 +392,9 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         CompoundNBT nbt = new CompoundNBT();
         nbt.put("fuelTank", fuelTank.serializeNBT());
         nbt.put("coolantTank", coolantTank.serializeNBT());
-        nbt.putFloat("fuelFertility", fuelFertility);
-        nbt.putFloat("fuelHeat", fuelHeat);
-        nbt.putFloat("reactorHeat", reactorHeat);
+        nbt.putDouble("fuelFertility", fuelFertility);
+        nbt.putDouble("fuelHeat", fuelHeat);
+        nbt.putDouble("reactorHeat", reactorHeat);
         return nbt;
     }
     
@@ -407,13 +407,13 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
             coolantTank.deserializeNBT(nbt.getCompound("coolantTank"));
         }
         if (nbt.contains("fuelFertility")) {
-            fuelFertility = nbt.getFloat("fuelFertility");
+            fuelFertility = nbt.getDouble("fuelFertility");
         }
         if (nbt.contains("fuelHeat")) {
-            fuelHeat = nbt.getFloat("fuelHeat");
+            fuelHeat = nbt.getDouble("fuelHeat");
         }
         if (nbt.contains("reactorHeat")) {
-            reactorHeat = nbt.getFloat("reactorHeat");
+            reactorHeat = nbt.getDouble("reactorHeat");
         }
     }
     
@@ -425,7 +425,7 @@ public class ClassicReactorSimulation implements INBTSerializable<CompoundNBT> {
         
         final int x;
         final int z;
-        float insertion = 0;
+        double insertion = 0;
         
         private ControlRod(int x, int z) {
             this.x = x;
